requirements.txt
beautifulsoup4==4.12.2
boto3==1.12.49
botocore==1.15.49
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
comtypes==1.1.10
coverage==7.3.1
crispy-bootstrap5==0.7
cryptography==41.0.4
decorator==5.1.0
digcommpy==0.9
distlib==0.3.7
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==3.2.20
django-appconf==1.0.5
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap-v5==1.0.11
django-cache-url==3.4.4
django-compressor==4.4
django-cors-headers==4.3.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-jquery==3.1.0
django-jsonfield==1.4.1
django-lgi==0.1.2
django-sass-processor==1.2.2
django-storages==1.14
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
docutils==0.15.2
Eel==0.12.4
environs==9.5.0
mysql-connector-python==8.1.0
nodeenv==1.8.0
Pillow==8.2.0
platformdirs==3.10.0
pooch==1.5.2
psycopg2-binary==2.9.7
PyMySQL==1.1.0
pyparsing==2.4.7
python-dotenv==1.0.0
requests==2.25.1
six==1.16.0
sqlparse==0.4.4
virtualenv==20.24.5
whitenoise==6.5.0
